requirements.txt
Flask==2.3.2
google-cloud-speech
google-generativeai
